requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
fabric3
awscli
boto3
GitPython
cryptography==2.4.2
